bug fixing with roles and post

+++ src/comments/comments.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Param, Post, UseGuards } from '@nestjs/common';
+import { Body, Controller, Param, Post, UnauthorizedException, UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { GetUser } from '../auth/get-user.decorator';
 import { User } from '../auth/user.entity';
@@ -14,16 +14,25 @@ export class CommentsController {
 
   @Post('/:id')
   createComment(@Body('commentString') commentString: string, @Param('id') postId: string, @GetUser() user: User) {
+    if(user.roles !== 'user'){
+      throw new UnauthorizedException('You dont have privileges to access this');
+    }
     return this.commentsService.createComment(commentString, postId, user);
   }
 
   @Post('/:id')
   deleteComment(@Param('id') commentId: string, user: User){
+    if(user.roles !== 'user'){
+      throw new UnauthorizedException('You dont have privileges to access this');
+    }
     return this.commentsService.deleteComment(commentId, user);
   }
 
   @Post('/:id')
   updateComment(@Body('commentString') commentString: string, @Param('id') commentId: string, @GetUser() user: User) {
+    if(user.roles !== 'user'){
+      throw new UnauthorizedException('You dont have privileges to access this');
+    }
     return this.commentsService.updateComment(commentString, commentId, user);
   }
